requirements.txt
Django==3.8.1
python-dotenv==1.1.0
Pillow==11.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-two-factor-auth==1.17.0
django-filter==25.1
